+++ Changes.md
@@ -1,6 +1,10 @@
 Version Information
 ===================
-Version 3.3.1.3.21
+Version 3.3.1.3.23
+  1. Stage 10 continued.
+  2. Fix section deleted (and therefore created too) - https://www.screencast.com/t/JAU2gBsfx.
+
+Version 3.3.1.3.22
   1. Stage 10.
   2. Fix drag and drop alignment - https://www.screencast.com/t/FqxZkac9L.